+++ src/WebRTC/RELEASE_NOTES.md
@@ -1,3 +1,14 @@
+### 1.0.0
+
+* First stable release
+* Fix `RTCIceCandidate.ip` as string option
+* Fix `RTCRtpSender.setStreams` should return a promise
+* Fix `RTCPeerConnection.setLocalDescription` parameter is mandatory
+* Fix `RTCPeerConnection.setRemoteDescription` parameter is mandatory
+* Fix `RTCPeerConnection.addIceCandidate` parameter is mandatory
+* Fix `RTCPeerConnection.addTrack` stream parameter is an array of `MediaStream`
+* Fix `RTCPeerConnection.addTransceiver` init parameter is type of type `RTCRtpTransceiverInit`
+
 ### 1.0.0-beta-001
 
 * First release